Moved classes out of redundant helper subpackage. Deleted help subpackage
