+++ src/main/java/edu/byu/ece/rapidSmith/device/Device.java
@@ -23,7 +23,7 @@ package edu.byu.ece.rapidSmith.device;
 import edu.byu.ece.rapidSmith.RSEnvironment;
 import edu.byu.ece.rapidSmith.design.PIP;
 import edu.byu.ece.rapidSmith.design.xdl.XdlPin;
-import edu.byu.ece.rapidSmith.device.helper.HashPool;
+import edu.byu.ece.rapidSmith.util.HashPool;
 import edu.byu.ece.rapidSmith.primitiveDefs.PrimitiveDefList;
 
 import java.io.Serializable;
