+++ src/main/java/edu/byu/ece/rapidSmith/device/creation/TileSources.java
@@ -18,7 +18,7 @@
  * get a copy of the license at <http://www.gnu.org/licenses/>.
  * 
  */
-package edu.byu.ece.rapidSmith.device.helper;
+package edu.byu.ece.rapidSmith.device.creation;
 
 import java.util.HashSet;
 import java.util.Set;
