--- src/main/java/edu/byu/ece/rapidSmith/device/helper/package.html
@@ -1,6 +1,0 @@
-<html><head></head><body>
-The device.helper package is a set of classes which are used for creating
-and populating the Device class and other classes in the device package.
-See <a href="http://rapidsmith.svn.sourceforge.net/viewvc/rapidsmith/trunk/doc/TechReportAndDocumentation.pdf">
-Technical Report and Documentation for more information. </a>
-</body></html>
